Enable continuous integration on Travis-CI
After every pull requests, tests will be executed and the results
displayed in GitHub interface. We could drop this after we integrate
with our Jenkins instance, or keep using Travis instead.

Travis runs the tests in a container based on Ubuntu. Some dependencies
are installed by APT, others from PyPI.

The problem here is that Kobo requires Koji and python-rpm packages,
which are not available and not easy to install. The hackish way to
resolve this is to modify kobo.rpmlib module not to include these
dependencies. This breaks it, but PDC does not need any of the broken
functions.

There is also a new settings file that disables the authentication. Make
target for test is updated to use these settings.