Log details about importing RPMs and images
The changeset includes details about compose, if it did not exist
before. It also includes a notice about total number of rpms/images that
were linked to that compose. This can be different to the number of
entities that were actually created since some might have already
existed.

A test for importing images is added.