Refactor compose/package
Add 2 new end points to take old compose/package
endpoint function.
One is  for a given release and srpm_name list all
composes that contain the package (and include its version).
Another is given a compose C and a package, find the
latest compose older than C which contains a
different version of the package.
Keep the old endpoint same for a while to don't break
existing scripts.

JIRA: PDC-645