Fix internal server error on OPTIONS requests
The metadata generator assumed that when a view responds to a PUT
request, it must be possible to call `get_object` on it to test object
permissions. This is however not the case with bulk operations, which
add PUT to list url. An unhandled AssertionError was raised in that
case.

This patch provides a custom BulkMetadata generator, which is pretty
much the same as the one provided by REST Framework, only it is aware
of the bulk operations.

The settings need to be updated in order for the generator to be used.
All settings files in this repository are updated, but anyone using
settings_local.py need to modify it accordingly.