Create compose-images resource
This is a resource that works pretty much the same as compose-rpms. It
allows import compose images and getting the same data back.

The old import API is still available, but marked as deprecated.

Since the existing code did not store path for the images (just the
filename), there are two migrations to get to the desired state. After
applying the migrations, all data will be preserved, only the previously
existing images will have an empty path.

JIRA: PDC-911