Provide list of filters automatically
We already have information about the available filters in
StrictQueryParamMixin. This patch gets the data and allows to use it for
automatic generation of documentation.

Depending on where the filter comes from various details about it are
available. It is possible to get lookup type and value type from a
FilterSet. For filters from filter_fields it would be difficult to get
the type (they are determined based on model). For extra_query_params
there is no general way to get any detail.

The filters are provided as Markdown formatted list. To use it from a
docstring, it is enough to place `%(FILTERS)s` where the list should be.

The list of filters for each API end-point is cached in memory.

There is a set of basic tests.