Enable automatically generated URLs in documentation
This patch modifies the browsable API renderer so that it is able to
replace certain strings in the docstrings with actual URLs obtained from
the router.

The general syntax for including urls is $TYPE:route-name:arg1:arg2$.
The number of arguments depends on the actual route, and there may be
none arguments required.

There are two ways types of links. $LINK:…$ will generate clickable link
with relative URL as a label. This should be used to show location of
different API end-points. Arguments will be wrapped in braces to clearly
show what user is supposed to fill in.

The other type is $URL:…$ which is replaced by an absolute URL pointing
to the specified place. This is useful for examples. Arguments are used
as specified, without any wrapping.

When the specified route can not be resolved, details are logged as an
error, and the rendered document includes BAD URL instead of the
link/URL.

The renderer is updated to work with Django REST Framework 3.2.2.