Remove hardcoded URLs from docs
This patch uses feature from previous commit to replace all hardcoded
URLs in documentation with automatically generated ones.

There are a couple of minor issues with some end-points:

After repos were removed, the save viewset is available under two
different names. The router seems to pick the last name, so order of
registrations in routers.py is important for this end-point..

The links for release variants API contain multiple parts, but to the
router these are specified as a single value (with the slash already
embedded into it). This slighly complicates proper wrapping of the
user-specified arguments.