Make sigkey key_id editable
This is consistent with other resources where their string primary
identifier can be changed as well (release_id, product_version_id etc).

This change makes it possible to allow PUT requests to update keys. The
original reason to disable it was because key_id was read-only.

A test for creating sigkeys is added to make sure it is not broken.

JIRA: PDC-634, PDC-855, PDC-965