Repo filters should accept duplicit filters
Since we expect to have a lot of repos, the filtering should allow
queries with multiple values to choose from (OR). Without this patch,
multiple occurrences of the same filter are accepted, but only the last
value is used for filtering.

Add a test suite for this feature. This is mostly to prevent regressions
in this feature.

JIRA: PDC-970