Separate auto doc test
Move test for auto documentation to a separate module that is not
included in make test. The reason for this is that the test hits most
end-points, but does not check much about them. This is raising the test
coverage significantly which could hide bugs.

The test is still run on Travis-CI and can be run locally via
    $ make extra_test

JIRA: PDC-975