Fix mangled filenames on partial update of RPM
When the request for partial update of RPM did not contain filename and
any of name, version, release and arch, the serializer would generate a
default filename which would be incorrect.

The actual way of creating the default is changed to only be invoked
when the filename is necessary. A couple of tests is added to make sure
this works correctly.

JIRA: PDC-959