Create new contact API
With this API, there is no inheritance between release and global
components. Instead, what would previously be inherited is stored
directly on a release component.

The relationship between a component and a contact is simplified to a
regular Many-to-Many relationship with an intermediate model to point to
role for that contact.

In this patch, new API is added and old tests are adapted to verify it
is correct. There is database schema migration, but the current data is
not migrated.

JIRA: PDC-1039