Find older compose from other releases
When requesting an older compose with different version of an RPM, the
composes from previous releases of the same product are explored as
well.

When going back, only GA releases are considered. This way, when
starting from compose for r-1.1-updates or r-1.2, composes from r-1.0
and r-1.1 will be considered, but not from r-1.0-updates.

    r-1.0 --------+- r-1.0-updates -----
                  |
                  \- r-1.1 -----+- r-1.1-updates -----
                                |
                                \- r-1.2 -----

JIRA: PDC-1008