Update client plugin infrastructure
The cli runner is moved to a separate module, so that tests can import
it and run it directly.

The makefile targets are updated:
 * cover_test runs coverage on server and client tests
 * extra_test is renamed to api_doc_test
 * client_test runs client tests
 * all_tests runs all tests

Part of the tests are based on a new test runner script. The client
tests do not run through the manage.py script.

The main part of this patch is the infrastracture for running client
tests. There is a base class that allows easy mocking of the API (with
no need to have a server running somewhere) and checking if the output
is correct.

There is also a helper script for creating new plugins and defining
tests for them. It saves the developer the hassle of creating
directories and boilerplate files.