Allow tracking OSBS metadata
Currently, there is only autorebuild flag. The information can be viewed
and modified through the API. The actual database records are created
and deleted automatically when component type changes or when the type
itself changes.

There is an internal change: the release component related field was
renamed so that the more useful one has an easy-to-use name.

JIRA: PDC-958