Fix new contact system
Main part of this patch is renaming new concepts so that the names are
as shorter but still clear to understand.

The filtering for the new resources is reworked to contain new filters
and remove duplicit old filters.

Documentation is added to new endpoints. All places where the old
contact system is visible in API are marked with a deprecation notice.
There are also comments in code to clearly mark which places need to be
edited when the old system is dropped.

Release cloning is updated to work with new contacts. As a side effect,
this fixed the issue about cloned contacts not being logged in
changesets.

The migrations are quite complex, and may need to be run with --noinput
to prevent a question about deleting stale content formats.

JIRA: PDC-1039