Return JSON 404 response from Django
When Django generates 404 response, it would always return the rendered
template, which is not nice as the client might be expecting JSON. This
patch adds a check for the Accept HTTP header. If it mentions
application/json MIME type, JSON object is returned instead.

This is automatically picked by pdc_client as it already adds the
header.

JIRA: PDC-1106