Improve performance of the web UI
Many of the pages performed lots of database queries. By selecting and
prefetching related data, that number can be brought down significantly.

This patch required rewriting the logic for determining if product or
product version is active. Now it  able to take advantage of the
preloaded data. A test case was added to make sure it works correctly.

JIRA: PDC-1132