Make release components be unique by name/release/type.
This used to consider only `name` and `release`, which makes a certain
amount of sense.  We don't want more than one rpm with the same name in
the same release entry here.

However, if we open this up to start handling other types of release
components like `container` or `module`, then it is easy to imagine that
we might have one "cockpit" rpm in a release but also a "cockpit"
container in the same release.

If you try to create both of those entries in PDC, it will fail saying
that "name and release must be unique".  This change fixes that by
loosening the uniqueness constraint to also take `type` into account.