Give a default value to ReleaseComponent.type in the Serializer.
This took a long time to figure out.

ReleaseComponent.type gets a default value in the model, which worked
fine up until 80397f1.  A `unique_together` constraint was added there.
Apparently, when you add the UniqueTogetherValidator into the mix, the
default values from the model don't propagate up to the serializer
anymore, so you have to specify an additional default value on the
serializer in order to retain previous behavior.

This gets the test suite working again.