Allow passing comma separated values to fields/exclude_fields
This is needed for PDC Contact Browser, which currently uses
"fields=email,mail_name".

It's interpreted as single field name ("email,mail_name") and doesn't
have any effect in PDC server 1.10.0 but in future versions this will be
treated as error.

So "fields" , and "exclude_fields" for consistency, shall support comma
separated list of fields (in string value).

JIRA: PDC-2656